backend/pyproject.toml
[project]
name = "gen-analytics"
version = "0.1.0"
description = "Natural language analytics tool for BigQuery"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-bigquery>=3.25.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.38.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.5.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
